+ thêm panel finish survival (thắng, thua)

+++ Assets/Scenes/Trung/Scritps/ScriptableObject/EnemySO.cs
@@ -27,7 +27,7 @@ public class EnemySO : ScriptableObject
 
     [Header("Rewards Player Can Receive")]
     public int Gems;
-    public float Exp;
+    public int Exp;
     public int Score = 10;
 
     private void OnValidate()
